delete by ajax for users,managers and coaches

+++ app/Http/Controllers/UserController.php
@@ -4,22 +4,46 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use App\Models\TrainingPackage;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use Illuminate\Support\Facades\Hash;
-
+use DataTables;
 
 
 class UserController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $users = User::all();
-        return view('tables.users', [
-            'users' => $users,
-        ]);
+        
+        if ($request->ajax()) {
+            $data = User::all();
+            return DataTables::of($data)
+                ->addIndexColumn()
+                ->addColumn('date', function ($row) {
+                    $data = $row->created_at->format('y-m-d');
+                    return $data;
+                })
+                ->addColumn('image', function ($row) {
+                    $src = asset('storage/images/'. $row->profile_image);
+                    return '<img src="' . $src . '" style="width:50px;height:50px;" />';
+                })
+                ->addColumn('action', function ($row) {
+                    return '<a href="' . route('users.edit', ['user' => $row->id]) . '" class="btn btn-primary">Edit</a>
+                    <button type="button" class="btn btn-danger " data-bs-toggle="modal"
+                    data-bs-target="#usermoadal"
+                    data-id=' . $row->id .'>
+                    delete
+                        </button>
+                        ';
+    
+                })
+                ->rawColumns(['image', 'action','date'])
+                ->make(true);
+        }
+        return view('tables.users');
     }
 
     public function create()
@@ -86,25 +110,24 @@ class UserController extends Controller
         return redirect()->route('users.index');
 
     }
-
     public function destroy($userId){
         $user = User::find($userId);
         if($user){
+            if($user->training_sessions()->count()){
+                return response()->json(["status"=>false,"message"=> "can't delete this user because still having sessions"] ,200);
+            }
             $deleted=$user->delete();
+              if($deleted){
+            return response()->json(["status"=>true,"message"=> "user no. ".$userId." deleted"] ,200);
+           }else{
+            return response()->json(["message"=> "something went wrong"] ,400);
+             }
         }else{
             return response()->json(["message"=> "could't find this user"] ,400);
         }
-        if($deleted){
-            return response()->json(["message"=> "user no. ".$userId." deleted"] ,200);
-        }else{
-            return response()->json(["message"=> "something went wrong"] ,400);
-        }
+      
         
         
-    }
-
-
-
-       
+    }       
     
 }
